Update OrganizationController.cs
Add validation

+++ WEB_API_Task_WebApplication/Controllers/OrganizationController.cs
@@ -31,7 +31,18 @@ namespace WEB_API_Task_WebApplication.Controllers
         [Route("Create")]
         [HttpPost]
         public List<Organization> Create(Organization item)
-        {
+        {       
+            if (item == null)
+            {
+                ModelState.AddModelError("", "No data for creating organization");
+                return BadRequest(ModelState);
+            }
+        
+             if (!ModelState.IsValid)
+             {
+                return BadRequest(ModelState);
+             }
+             
             _organizationRepository.Create(item);
 
             return _organizationRepository.Entities;
@@ -41,6 +52,18 @@ namespace WEB_API_Task_WebApplication.Controllers
         [HttpPost]
         public List<Organization> Delete(int Id)
         {
+  
+        if (Id == null)
+            {
+                ModelState.AddModelError("", "No data for deleting organization");
+                return BadRequest(ModelState);
+            }
+        
+             if (!ModelState.IsValid)
+             {
+                return BadRequest(ModelState);
+             }
+            
             _organizationRepository.Delete(Id);
 
             return _organizationRepository.Entities;
@@ -50,6 +73,18 @@ namespace WEB_API_Task_WebApplication.Controllers
         [HttpPost]
         public List<Organization> Update(Organization item) 
         {
+        
+         if (item == null)
+            {
+                ModelState.AddModelError("", "No data for updating organization");
+                return BadRequest(ModelState);
+            }
+        
+             if (!ModelState.IsValid)
+             {
+                return BadRequest(ModelState);
+             }
+         
            _organizationRepository.Update(item);
 
             return _organizationRepository.Entities;
